Remove duplicated code in new files
