+++ client/render/CBitmapHanFont.cpp
@@ -18,273 +18,6 @@
 #include "../../lib/filesystem/Filesystem.h"
 #include "../../lib/CGeneralTextHandler.h"
 
-size_t IFont::getStringWidth(const std::string & data) const
-{
-	size_t width = 0;
-
-	for(size_t i=0; i<data.size(); i += Unicode::getCharacterSize(data[i]))
-	{
-		width += getGlyphWidth(data.data() + i);
-	}
-	return width;
-}
-
-void IFont::renderTextLeft(SDL_Surface * surface, const std::string & data, const SDL_Color & color, const Point & pos) const
-{
-	renderText(surface, data, color, pos);
-}
-
-void IFont::renderTextRight(SDL_Surface * surface, const std::string & data, const SDL_Color & color, const Point & pos) const
-{
-	Point size((int)getStringWidth(data), (int)getLineHeight());
-	renderText(surface, data, color, pos - size);
-}
-
-void IFont::renderTextCenter(SDL_Surface * surface, const std::string & data, const SDL_Color & color, const Point & pos) const
-{
-	Point size((int)getStringWidth(data), (int)getLineHeight());
-	renderText(surface, data, color, pos - size / 2);
-}
-
-void IFont::renderTextLinesLeft(SDL_Surface * surface, const std::vector<std::string> & data, const SDL_Color & color, const Point & pos) const
-{
-	Point currPos = pos;
-
-	for(const std::string & line : data)
-	{
-		renderTextLeft(surface, line, color, currPos);
-		currPos.y += (int)getLineHeight();
-	}
-}
-
-void IFont::renderTextLinesRight(SDL_Surface * surface, const std::vector<std::string> & data, const SDL_Color & color, const Point & pos) const
-{
-	Point currPos = pos;
-	currPos.y -= (int)data.size() * (int)getLineHeight();
-
-	for(const std::string & line : data)
-	{
-		renderTextRight(surface, line, color, currPos);
-		currPos.y += (int)getLineHeight();
-	}
-}
-
-void IFont::renderTextLinesCenter(SDL_Surface * surface, const std::vector<std::string> & data, const SDL_Color & color, const Point & pos) const
-{
-	Point currPos = pos;
-	currPos.y -= (int)data.size() * (int)getLineHeight() / 2;
-
-	for(const std::string & line : data)
-	{
-		renderTextCenter(surface, line, color, currPos);
-		currPos.y += (int)getLineHeight();
-	}
-}
-
-std::array<CBitmapFont::BitmapChar, CBitmapFont::totalChars> CBitmapFont::loadChars() const
-{
-	std::array<BitmapChar, totalChars> ret;
-
-	size_t offset = 32;
-
-	for (auto & elem : ret)
-	{
-		elem.leftOffset =  read_le_u32(data.first.get() + offset); offset+=4;
-		elem.width =       read_le_u32(data.first.get() + offset); offset+=4;
-		elem.rightOffset = read_le_u32(data.first.get() + offset); offset+=4;
-	}
-
-	for (auto & elem : ret)
-	{
-		int pixelOffset =  read_le_u32(data.first.get() + offset); offset+=4;
-		elem.pixels = data.first.get() + 4128 + pixelOffset;
-
-		assert(pixelOffset + 4128 < data.second);
-	}
-	return ret;
-}
-
-CBitmapFont::CBitmapFont(const std::string & filename):
-	data(CResourceHandler::get()->load(ResourceID("data/" + filename, EResType::BMP_FONT))->readAll()),
-	chars(loadChars()),
-	height(data.first.get()[5])
-{}
-
-size_t CBitmapFont::getLineHeight() const
-{
-	return height;
-}
-
-size_t CBitmapFont::getGlyphWidth(const char * data) const
-{
-	std::string localChar = Unicode::fromUnicode(std::string(data, Unicode::getCharacterSize(data[0])));
-
-	if (localChar.size() == 1)
-	{
-		const BitmapChar & ch = chars[ui8(localChar[0])];
-		return ch.leftOffset + ch.width + ch.rightOffset;
-	}
-	return 0;
-}
-
-void CBitmapFont::renderCharacter(SDL_Surface * surface, const BitmapChar & character, const SDL_Color & color, int &posX, int &posY) const
-{
-	Rect clipRect;
-	CSDL_Ext::getClipRect(surface, clipRect);
-
-	posX += character.leftOffset;
-
-	CSDL_Ext::TColorPutter colorPutter = CSDL_Ext::getPutterFor(surface, 0);
-
-	uint8_t bpp = surface->format->BytesPerPixel;
-
-	// start of line, may differ from 0 due to end of surface or clipped surface
-	int lineBegin = std::max<int>(0, clipRect.y - posY);
-	int lineEnd   = std::min<int>(height, clipRect.y + clipRect.h - posY - 1);
-
-	// start end end of each row, may differ from 0
-	int rowBegin = std::max<int>(0, clipRect.x - posX);
-	int rowEnd   = std::min<int>(character.width, clipRect.x + clipRect.w - posX - 1);
-
-	//for each line in symbol
-	for(int dy = lineBegin; dy <lineEnd; dy++)
-	{
-		uint8_t *dstLine = (uint8_t*)surface->pixels;
-		uint8_t *srcLine = character.pixels;
-
-		// shift source\destination pixels to current position
-		dstLine += (posY+dy) * surface->pitch + posX * bpp;
-		srcLine += dy * character.width;
-
-		//for each column in line
-		for(int dx = rowBegin; dx < rowEnd; dx++)
-		{
-			uint8_t* dstPixel = dstLine + dx*bpp;
-			switch(srcLine[dx])
-			{
-			case 1: //black "shadow"
-				colorPutter(dstPixel, 0, 0, 0);
-				break;
-			case 255: //text colour
-				colorPutter(dstPixel, color.r, color.g, color.b);
-				break;
-			default :
-				break; //transparency
-			}
-		}
-	}
-	posX += character.width;
-	posX += character.rightOffset;
-}
-
-void CBitmapFont::renderText(SDL_Surface * surface, const std::string & data, const SDL_Color & color, const Point & pos) const
-{
-	if (data.empty())
-		return;
-
-	assert(surface);
-
-	int posX = pos.x;
-	int posY = pos.y;
-
-	// Should be used to detect incorrect text parsing. Disabled right now due to some old UI code (mostly pregame and battles)
-	//assert(data[0] != '{');
-	//assert(data[data.size()-1] != '}');
-
-	SDL_LockSurface(surface);
-
-	for(size_t i=0; i<data.size(); i += Unicode::getCharacterSize(data[i]))
-	{
-		std::string localChar = Unicode::fromUnicode(data.substr(i, Unicode::getCharacterSize(data[i])));
-
-		if (localChar.size() == 1)
-			renderCharacter(surface, chars[ui8(localChar[0])], color, posX, posY);
-	}
-	SDL_UnlockSurface(surface);
-}
-
-std::pair<std::unique_ptr<ui8[]>, ui64> CTrueTypeFont::loadData(const JsonNode & config)
-{
-	std::string filename = "Data/" + config["file"].String();
-	return CResourceHandler::get()->load(ResourceID(filename, EResType::TTF_FONT))->readAll();
-}
-
-TTF_Font * CTrueTypeFont::loadFont(const JsonNode &config)
-{
-	int pointSize = static_cast<int>(config["size"].Float());
-
-	if(!TTF_WasInit() && TTF_Init()==-1)
-		throw std::runtime_error(std::string("Failed to initialize true type support: ") + TTF_GetError() + "\n");
-
-	return TTF_OpenFontRW(SDL_RWFromConstMem(data.first.get(), (int)data.second), 1, pointSize);
-}
-
-int CTrueTypeFont::getFontStyle(const JsonNode &config)
-{
-	const JsonVector & names = config["style"].Vector();
-	int ret = 0;
-	for(const JsonNode & node : names)
-	{
-		if (node.String() == "bold")
-			ret |= TTF_STYLE_BOLD;
-		else if (node.String() == "italic")
-			ret |= TTF_STYLE_ITALIC;
-	}
-	return ret;
-}
-
-CTrueTypeFont::CTrueTypeFont(const JsonNode & fontConfig):
-	data(loadData(fontConfig)),
-	font(loadFont(fontConfig), TTF_CloseFont),
-	blended(fontConfig["blend"].Bool())
-{
-	assert(font);
-
-	TTF_SetFontStyle(font.get(), getFontStyle(fontConfig));
-}
-
-size_t CTrueTypeFont::getLineHeight() const
-{
-	return TTF_FontHeight(font.get());
-}
-
-size_t CTrueTypeFont::getGlyphWidth(const char *data) const
-{
-	return getStringWidth(std::string(data, Unicode::getCharacterSize(*data)));
-	/*
-	int advance;
-	TTF_GlyphMetrics(font.get(), *data, nullptr, nullptr, nullptr, nullptr, &advance);
-	return advance;
-	*/
-}
-
-size_t CTrueTypeFont::getStringWidth(const std::string & data) const
-{
-	int width;
-	TTF_SizeUTF8(font.get(), data.c_str(), &width, nullptr);
-	return width;
-}
-
-void CTrueTypeFont::renderText(SDL_Surface * surface, const std::string & data, const SDL_Color & color, const Point & pos) const
-{
-	if (color.r != 0 && color.g != 0 && color.b != 0) // not black - add shadow
-		renderText(surface, data, Colors::BLACK, pos + Point(1,1));
-
-	if (!data.empty())
-	{
-		SDL_Surface * rendered;
-		if (blended)
-			rendered = TTF_RenderUTF8_Blended(font.get(), data.c_str(), color);
-		else
-			rendered = TTF_RenderUTF8_Solid(font.get(), data.c_str(), color);
-
-		assert(rendered);
-
-		CSDL_Ext::blitSurface(rendered, surface, pos);
-		SDL_FreeSurface(rendered);
-	}
-}
-
 size_t CBitmapHanFont::getCharacterDataOffset(size_t index) const
 {
 	size_t rowSize  = (size + 7) / 8; // 1 bit per pixel, rounded up
